### 2025-04-16 #7913 @goldingn
@tzuV can you please suggest some potential reviewers? Ideally people who have some experience with research software and with simulacrum or similar methods? They don't have to be R experts (we can find other reviewers for that), though if they are that's even better.

Please mention their github handles here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
