### 2025-04-22 #7913 @goldingn
@agranholm would you be interested in reviewing this R package submission on interacting with a clinical database for JOSS?
