### 2025-04-22 #7913 @PatrickRWright
> [@PatrickRWright](https://github.com/PatrickRWright) would you be interested in reviewing this R package submission on interacting with a clinical database for JOSS?

Sorry I can't do it. Maybe @DrEspresso or @aghaynes are interested.
