### 2025-04-22 #7913 @agranholm
> [@agranholm](https://github.com/agranholm) would you be interested in reviewing this R package submission on interacting with a clinical database for JOSS?

@goldingn : Thanks for the request, but this is unfortunately somewhat outside my area of expertise.
