### 2025-04-24 #7913 @aghaynes
Hi @goldingn

I'm not familiar with simulacrum at all i'm afraid, but I could perhaps review it from the R users side if thats useful?

I had a VERY quick look at the package and noticed one potential place to make it easier for the R user... the user currently needs to write SQL code to perform a query. Maybe it's possible to lever e.g. dbplyr so that the user can write R code instead and have dbplyr return the necessary SQL code...? That might be a substantial task to implement though...
