### 2025-04-24 #7913 @tzuV
Hi @aghaynes, thank you for the valuable suggestion.

I've investigated dbplyr and believe it's feasible to integrate it into the package within a reasonable timeframe. The ability to translate dplyr pipelines into SQL via show_query() aligns well with the package's context.

I think an R user's review would be great. Building a user-friendly system is crucial, especially for those familiar with Simulacrum who need streamlined tools to access the real cancer database, as discussed further in the paper.
