### 2025-04-28 #7913 @goldingn
@aghaynes Thank you! there's no need to be an expert in the specific dataset, your R experience and general knowledge of clinical data and database interactions from R are exactly what we need.

Now we have 2 reviewers, I'll add you both and kick-off the review process. Please see instructions in the new review issue for your checklists and next steps.

@DrEspresso whilst we now have 2 reviewers, you'd still be welcome to join as a third reviewer if you are interested. Just let me know if so.
