### 2025-04-28 #8120 @editorialbot
[REVIEW]: simulacrumWorkflowR: An R package for Streamlined Access and Analysis of the Simulacrum Cancer Dataset
**Submitting author:** @tzuV (<a href="http://orcid.org/0009-0006-8967-9268">Jakob Skelmose</a>)
**Repository:** https://github.com/CLINDA-AAU/simulacrumWorkflowR.git
**Branch with paper.md** (empty if default branch): main
**Version:** v3
