### 2025-04-29 CLINDA-AAU/simulacrumWorkflowR#1 @aghaynes
JOSS Review
As I work through the review, I'll note anything I see in this issue (I won't do everything all at once, so this is just a start)...

Line 49 of the paper (compiled PDF) seems to be missing a reference?
Line 81 of the paper (installation R code) seems to have an additional closing `)`

Maybe i missed it in the paper, but including the link to download Simulacrum would be useful (i know it's on your readme)

The limitations section looks like it should be a list, but isn't formatted as such in the PDF... maybe there's a syntax error?

While it looks like you have some example code in the helpfiles, you don't really have any testing implemented. I would encourage you to implement some tests with the `testthat` package to ensure that when you change your code, you don't break things and that your outputs are as expected. (Note that one review point is `Are there automated tests or manual steps described so that the functionality of the software can be verified?`)

#### Other things that would be nice additions to your package, but aren't requirements for the JOSS paper

A thought... maybe a function that opens a web browser to the relevant simulacrum download page would be handy...?

I suggested elsewhere that having the create_workflow function parse e.g. dbplyr syntax to the required SQL would aid the R user a lot

I would suggest implementing a pkgdown website. It's really easy to do, even to have it automatically update when you update your repo. (I could make a PR with the necessary files if you like)

I have to stop for today... more soon (or not, we'll see :wink:)
