### 2025-04-29 CLINDA-AAU/simulacrumWorkflowR#1 @tushardave26
@aghaynes - Thanks for starting this. I will add my review comments here, as they overlap with yours and are relatively few.

@tzuV - I agree with all that @aghaynes said about the typo at line 81 of the paper and the possibly missing reference at line 49. I would also suggest you add automated tests using the `testthat` package to adhere to the JOSS guidelines. I also noticed that the repo is missing a contributing guide (**CONTRIBUTING.md**) for others. Please review the reviewer's checklist in the review [thread](https://github.com/openjournals/joss-reviews/issues/8120), which addresses the **Community Guidelines**.
