### 2025-04-30 CLINDA-AAU/simulacrumWorkflowR#1 @aghaynes
The example code in the paper doesnt seem to work...

```r
> library(simulacrumWorkflowR)
> Dir <- "simulacrum/simulacrum_v2.1.0/Data"
> simulacrum_list_df <- read_simulacrum(Dir)
  |                                                  | 0 % ~calculating  Reading: sim_av_gene
  |+++++                                             | 9 % ~31s          Reading: sim_av_patient
  |++++++++++                                        | 18% ~01m 04s      Reading: sim_av_tumour
  |++++++++++++++                                    | 27% ~01m 35s      Reading: sim_rtds_combined
  |+++++++++++++++++++                               | 36% ~03m 56s      Reading: sim_rtds_episode
  |+++++++++++++++++++++++                           | 45% ~03m 04s      Reading: sim_rtds_exposure
  |++++++++++++++++++++++++++++                      | 55% ~02m 56s      Reading: sim_rtds_prescription
  |++++++++++++++++++++++++++++++++                  | 64% ~02m 18s      Reading: sim_sact_cycle
  |+++++++++++++++++++++++++++++++++++++             | 73% ~01m 33s      Reading: sim_sact_drug_detail
  |+++++++++++++++++++++++++++++++++++++++++         | 82% ~60s          Reading: sim_sact_outcome
  |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~27s          Reading: sim_sact_regimen
  |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=04m 35s
Files successfully loaded!
Warning message:
In read_simulacrum(Dir) :
  Please refer to tables by their original names, capitalized as presented (e.g., SIM_AV_PATIENT)
> query <- “SELECT *
Error: unexpected invalid token in "query <- “"
> query <- "SELECT *
+     FROM sim_av_patient
+ INNER JOIN sim_av_tumour ON sim_av_patient.patientid = sim_av_tumour.patientid"
> merged_data <- query_sql(query)
Error: no such table: sim_av_patient
In addition: Warning message:
package ‘RSQLite’ was built under R version 4.4.3
```

I think the problem is that you need to assign the tables from the list to the global environment:

```r
sim_av_patient <- simulacrum_list_df$sim_av_patient
sim_av_tumour<- simulacrum_list_df$sim_av_tumour
```

Line 84, the example doesn't work - using the create_workflow function: that function doesn't have a `Logger_report` argument... I guess you mean `logging`...

I think some further documentation (or links to where there is already information) on writing the commands in `create_workflow` would be useful. Looking at the workflow generated by your example code, I don't see how it would work on the CAS servers (I may well be wrong though)... For example, `cancer_grouping` is used with argument `sim_av_tumour`, which as far as I can tell from the code doesn't exist (it's output is also not assigned to anything). Furthermore, you use `html_table_model()`, but i'm unaware of any such function (maybe thats a function in CAS though?)...

```r
# Query ----------------------------------------------------------
query1 <- "SELECT * FROM sim_av_tumour WHERE age > 50 AND ROWNUM <= 500;"
data <- dbGetQuery(my_oracle, query1)
# Data Management ----------------------------------------------------------
cancer_grouping(sim_av_tumour)
# Analysis ----------------------------------------------------------
model = glm(AGE ~ STAGE_BEST + GRADE, data=data)
# Model Results ----------------------------------------------------------
html_table_model(model)
```

Your example code in the helpfile to `create_workflow` is also likely to be invalid (you cannot run write.csv on a glm object...). Please use code that actually works in helpfiles... (yes, the example will run, but the workflow file it generates will/may not)...

```r
create_workflow(
  libraries = "library(dplyr)",
  query = "SELECT * FROM sim_av_patients LIMIT age > 50;",
  data_management = "df <- cancer_grouping(df)",
  analysis = "model <- glm(Y ~ x1 + x2 + x3, data=df",
  model_results = "write.csv(model, 'results.csv')"
)
```

I noticed that you have a `hello` function in your package, which looks like it's from the package template... maybe you want to remove that... 😉
