### 2025-04-30 CLINDA-AAU/simulacrumWorkflowR#1 @tushardave26
Thank you for your thoughtful feedback, @aghaynes. I’d like to share a few observations that might help clarify things.

From what I’ve seen, the examples in the **README** appear to be the most up-to-date and reliable. I tried them out yesterday, and everything worked as expected. In contrast, the examples in the paper seem more illustrative in nature rather than intended for direct execution, although I may be mistaken. I’ll defer to @tzuV to clarify the intended use and purpose of those code snippets in the manuscript.

If the examples in the paper are meant to be functional, then I agree with your point, they need to be reviewed and corrected. In the meantime, I encourage trying the examples in the **README**, as they provide a more accurate and reproducible path through the tool's core functionality and should address most of your concerns.

To help avoid this kind of confusion going forward, I would suggest that @tzuV consider **maintaining consistency** between the examples in the paper and those in the **README**.
