### 2025-05-01 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Thank you @aghaynes and @tushardave26 for your comments.

I will try to cover as much as possible today and tomorrow.

I agree with the code suggestions from @aghaynes. I think they are valuable and will benefit the package, and I will add them all. Thank you for offering to provide a PR, @aghaynes! If it is not too much trouble and aligns with the reviewer guidelines, I would greatly appreciate your contribution :)

I will also correct the typos and syntax errors, remove the `hello` function, add a **CONTRIBUTING.md**, and add the testthat package to the package.

Regarding the example in the paper, yes, it was more of an illustrative example rather than a functional example, as @tushardave26 noted. I realize it would be more convenient to streamline the example code in the README.md and the paper's code for better consistency and usability.

Please correct me if I missed something that needs to be addressed.
