### 2025-05-01 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
To-Do List for R Package Changes.

Paper Changes:

[X] Check Line 49 of the compiled PDF and add the missing reference.
[ ] Correct the typo on Line 81 of the compiled PDF (installation R code) by removing the additional closing parenthesis ')'.
[ ] Add a link to the Simulacrum download page within the paper.
[X] Reformat the Limitations section in the paper to be a proper list (addressing potential syntax errors causing incorrect formatting).
[ ] Review and correct the code examples in the paper to ensure they are functional and consistent with the examples provided in the README.md file.

R Package Code & Structure Changes:

[ ] Implement automated tests using the testthat package.
[ ] Remove the hello function.
[ ] Implement a function that opens a web browser to the relevant Simulacrum download page.
[ ] Explore implementing functionality in the create_workflow function to parse dbplyr syntax into the required SQL.

Package Documentation & Files:

[ ] Add a CONTRIBUTING.md file to the repository.
[ ] Implement a pkgdown website for the package.
