### 2025-05-01 CLINDA-AAU/simulacrumWorkflowR#1 @aghaynes
Seems that making the pkgdown site is a bigger job than i thought, primarily because of the use of local copies of the full simulacrum dataset in vignettes, which stops pkgdown compiling the pages.

One solution would be to include a tiny subset (e.g. data from 10 patients) in the package... You should probably check the licensing on the data though, to see if it's allowed. Selection of the patients might also be tricky to ensure that your vignette examples all still work.

If it's possible, this would also allow you to run some testthat tests automatically, as you'd have data to test against.
