### 2025-05-01 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Thank you for investigating these matters and for your advice, @aghaynes.

I suggest creating a small dataset that mirrors a subset of the Simulacrum, with the cell values generated randomly from a distribution.
