### 2025-05-02 #7913 @DrEspresso
@goldingn Thanks for considering me as reviewer for this paper! Unfortunately I'm late to the party, but I would be happy to contribute to the review, if @aghaynes and @tushardave26 are not too much ahead.
