### 2025-05-03 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
@aghaynes, I have created dummy data for the unit testing, which I am almost done with, and made the structure in the repo, as you described.

I have implemented a `browseURL` function to automatically open the data request site on Simulacrum (which was a nice addition to the package).

I have worked intensively on trying to implement the `dbplyr` package in a sensible way into the package. I struggle with finding a strategy for implementing `dbplyr` in a way that works and does not break the package. I have added an example in `vignette/simulacrumWorkflowR.Rmd` on how to use `dbplyr` to build queries in the workflow. I hope it is fine for now. I will continue thinking about other solutions, while also being open to ideas.

I will push my latest changes, and I hope it is fit for a `pkgdown` page now; otherwise, please let me know.
