### 2025-05-05 CLINDA-AAU/simulacrumWorkflowR#1 @aghaynes
Thanks @tsuV. For the dummy data, i would suggest that it mimics the same structure as the actual simulacrum so that you can use exactly the same code on it... For instance, your read_simulacrum function checks for the names of the different datasets, so replacing the examples in the vignettes doesnt work (even after renaming the two datasets you created to their equivalent names)...
```r
dir <- system.file("extdata/minisimulacrum", package = "simulacrumWorkflowR")
data_frames_lists <- read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour"))
Error in `read_simulacrum()`:
! Missing required files: sim_av_gene.csv, sim_rtds_combined.csv, sim_rtds_episode.csv, sim_rtds_exposure.csv, sim_rtds_prescription.csv, sim_sact_cycle.csv, sim_sact_drug_detail.csv, sim_sact_outcome.csv, sim_sact_regimen.csv
```
Maybe it's better/easier to sample from the actual data instead? I could imagine selecting a set of patients, then shuffling the data in each variable...
