### 2025-05-05 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Yes, I can see what you mean @aghaynes. I will try to make some new dummy data based on the simulacrum, but with shuffled values in each cell.
