### 2025-05-05 CLINDA-AAU/simulacrumWorkflowR#1 @tushardave26
@tzuV -- Thanks for making the changes rapidly. I would second @aghaynes' suggestion for the dummy data. When working with a third-party data source, the best way to create the dummy data is to select random data from the original data. Another thought is to remove identifying info for the patients, as now you are generating the dummy data from the real data. I am sure the real data must contain the de-identified data, but it's worth to double-check that.
