+++ vignettes/simulacrumWorkflowR.Rmd
@@ -32,7 +32,7 @@ library(dbplyr)
 
 # Read the dataset
 ```{r}
-dir <- "C:/Users/p90j/Documents/simulacrum_v2.1.0/Data"
+dir <- system.file("extdata/minisimulacrum", package = "simulacrumWorkflowR")
 data_frames_lists <- read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour")) 
 ```
 
