### 2025-05-06 7914201
updated unit testing according to the new dummy data
