+++ NAMESPACE
@@ -14,6 +14,7 @@ export(sqlite2oracle)
 export(start_time)
 export(stop_time)
 export(survival_days)
+import(dplyr)
 importFrom(DBI,dbConnect)
 importFrom(DBI,dbDisconnect)
 importFrom(DBI,dbIsValid)
