### 2025-05-06 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Thank you for the suggestions @aghaynes and @tushardave26. I have now created additional dataframes to make a complete selection of dummy dataframes, which can be used for running the vignette and the unit testing. However, I created a data generation function that created dataframes that matched the structure of the Simulacrum dataframes, but inserted constants instead of shuffling the values. As an exception, I have engineered the datasets "...patients" and "...tumour" to fit the code in the vignette. I am cautious in the generation of dummy data. I have tried to design a solution where there is no doubt that the data cannot be traced to the real data, while also making sure the data functions for demonstration and testing in the package. If we need to modify the dummy data more, please let me know.
