### 2025-05-07 #7913 @goldingn
Hi @DrEspresso, that's great! You'd be very welcome.

They have made some headway on the [main review thread here](https://github.com/openjournals/joss-reviews/issues/8120) and [with detailed discussion here](https://github.com/CLINDA-AAU/simulacrumWorkflowR/issues/1))

I'll add you as a reviewer over on the main review thread now, you'll then need to ask editorialbot to create a checklist for you and you can get stuck in.
