### 2025-05-16 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
@tzuV JOSS requires that the repository contains a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license. One option would be to use a licencing function of the [usethis](https://usethis.r-lib.org/reference/licenses.html) package to create a LICENSE.md file in accordance with CRAN requirements.
