### 2025-05-18 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Hi @DrEspresso, thank you for joining as a reviewer for this paper and for the tip regarding the `usethis` package.  I have chosen the MIT license, as it seemed most appropriate for this package, and I have pushed the changes to the repository.
