### 2025-05-20 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
Hi @tzuV, thank you for adding the LICENSE.md. I enjoyed reading your paper, and I think your package is a valuable tool to streamline data processing with the CAS data. I agree to the comments of @aghaynes and @tushardave26, and thank you for implementing their feedback! While reading your paper, I found a few points you may want to address in order to improve the readability further:
- You could add a reference / link to Simulacrum, Cancer Analysis System, HDI. I guess from the Acknowledgements, HDI is short for Health Data Insight? I would add this to the main text. I also suggest adding regional labels, e.g. United Kingdom's National Disease Registration Service, or similar.
- In the summary, the workflow is described as "accessing, preprocessing, and validating statistical analyses on the Simulacrum data set". I found the term "validating" misleading, as the package does not evaluate analyses. I suggest to use "testing" and/or "planning" instead.
- In line 30 you write "releasing the data". To my understanding, the data remain with NDRS, and only aggregated results are released? Similarly, in line 43 you write "accessing the real data through Simulacrum".
- As the limit of 3 hours free of charge may change in the future, I suggest to add a "as of 2025" to the sentence.
- In figure 1 I suggest to remove the version of Simulacrum, unless this version is a requirement. If so, then please add this to the Limitations and mention it in the main text. Also, I suggest swapping "NHS" for "NDRS".
- In line 70 you could swap "NHS servers" with "CAS database servers" (for consistency with the introduction).
- Please cite R and the R packages from the dependencies. See e.g. https://ropensci.org/blog/2021/11/16/how-to-cite-r-and-r-packages/
- In the code snippet for the installation, I recommend to remove `if(!require("devtools"))`, and add a comment `#` before `install.packages("devtools")`
- In the code snippet for querying the data, I suggest to add a linebreak before ON to avoid the line beeing cut-of when printing the paper.
- The described limitations are generic to the Simulacrum data, and not specific to the package. You could mention this.
- Data coverage: The 2016 - 2019 coverage is speficic to version v2.1.0, but your package should also work with the earlier version v1.2.0? Can you make this more generic? Also, I don't understand why this restriction needs to be applied to the code before running on CAS data? I assume a reseacher will be happy to draft and test the query on Simulacrum on the restricted data set, but will ask NDRS to run the query on the full data available? Please explain if there are additional limitations or other reasons why the restriction on CAS would be recommended.
- Structure: Is there a reference where a researcher can look up the structure of the CAS database and try to adjust the query themselves?
- In line 96 a linebreak is likely missing before "Time Management".
- I think you could add a reference to the CAS data: https://doi.org/10.1093/ije/dyz076
- Finally, I recommend to consistently use "the Simulacrum" or "Simulacrum".
