### 2025-05-20 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
I think the repository would benefit from adding bug report and feature request issue templates: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository
