### 2025-05-20 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
I don't think the output of `compute_time_limit()` is very useful. The [Submitting Code to NDRS guide](https://digital.nhs.uk/ndrs/data/data-outputs/cancer-publications-and-tools/simulacrum/simulacrum-user-guide/submitting-code-to-ndrs-for-a-data-release) states

> If it takes less than 3 hours of a NDRS analyst’s time to process the request [...], the work can be done free of charge ...

As noted in the function's warning, the actual time needed by the analyst has to be taken into account. I don't know if the actual computing time is relevant for the NDRS. If the analyst sets up the workflow and submits the script in one hour, the script runs then for two hours on the database, followed by the analyst checking the output and sending the results in another hour: Does the overall time period of 4 hours count, or the effective time of the analyst (2 hours)?

Under these circumstances, and given the different processing powers, I would advise against categorizing the queries in "accepted" and "rejected", and leave the interpretation of the total execution time to the user.
