### 2025-05-20 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
Installing the package works, and I could run through the vignette with the minimal dataset, thank you for this addition!
However, I get the following warning:

```
Warning message:
In read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour")) :
  Please refer to tables by their original names, capitalized as presented (e.g., SIM_AV_PATIENT)
```

Perhaps you want to change to an example without warning, or explain in the vignette why the warning appears.

Finally, I think the examples of `create_workflow()` in the paper, the README, and the vignette, create invalid workflows and I recommend to carefully review them and edit as necessary.

- According to the [Simulacrum user guide](https://simulacrum.healthdatainsight.org.uk/wp-content/uploads/2023/04/Simulacrum-v2-User-Guide.pdf) the table names between CAS and Simulacrum differ. Ideally, create_workflow() would replace the Simulacrum table names with CAS table names, e.g. AV_PATIENT instead of SIM_AV_PATIENT.

- In the vignette `create_workflow()` results in an R script where the output of query1 is returned as data frame "data" (which I would suggest to change to "query_result"), but in the data management section it is assumed the output is named "df1":

```
query1 <- "SELECT *
FROM sim_av_patient
INNER JOIN sim_av_tumour ON sim_av_patient.patientid = sim_av_tumour.patientid;"
data <- dbGetQuery(my_oracle, query1)
# Data Management ----------------------------------------------------------
df2 <- survival_days(df1)
```

- In the README, `cancer_grouping(sim_av_tumour)` is placed in the data management section, but `sim_av_tumour` is not defined and the output is printed instead of assigned to an object:

```
query1 <- "SELECT *
FROM SIM_AV_PATIENT
WHERE age > 50
AND ROWNUM <= 500;"
data <- dbGetQuery(my_oracle, query1)
# Data Management ----------------------------------------------------------
# Run query on SQLite database
cancer_grouping(sim_av_tumour)
# Additional preprocessing
#df2 <- survival_days(df1)
# Analysis ----------------------------------------------------------
model = glm(AGE ~ STAGE_BEST + GRADE,  data=data)
```

- Finally, in the vignette, the analysis refers to a variable GENDER.X, which does not exist, because SQL will (unlike R's dbplyr) not append .x and .y to the variable names present in both tables. If GENDER is indeed present in both tables, I suggest to specifiy which column to use in the SELECT statement and work with aliases, e.g. "SELECT SIM_AV_PATIENT.GENDER gender, ..."

These are examples I noticed, but not necessarily all cases where a workflow result could fail on the CAS database.
