### 2025-05-20 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
> **R Package Code & Structure Changes:**
>
> [X] Implement a function that opens a web browser to the relevant Simulacrum download page.

Thank you for picking up this idea. I suggest to add the URL to Simulacrum's download page nevertheless in the documentation. I run R on WSL, so `open_simulacrum_request()` does not work for me. It would be handy if the URL could be checked without opening the source code.
