### 2025-05-22 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
**Paper/Documentation Content & Clarity**
[ ] Add references/links to Simulacrum, Cancer Analysis System (CAS), and Health Data Insight (HDI).

[ ] Clarify "HDI" as "Health Data Insight" in the main text and add regional labels (e.g., "United Kingdom's National Disease Registration Service").

[ ] In the summary, change "validating statistical analyses" to "testing" and/or "planning" statistical analyses.

[ ] In line 30 & 43, clarify "releasing the data" to reflect that aggregated results are released, not the raw data.

[ ] Add "as of 2025" to the sentence regarding the 3-hour free-of-charge limit.

[ ] In Figure 1: Remove the Simulacrum version unless it's a requirement (if so, add to Limitations and main text).

[ ] Swap "NHS" for "NDRS".

[ ] In line 70, swap "NHS servers" with "CAS database servers".

[ ] Cite R and all R packages from the dependencies (e.g., using citation() in R).

[ ] Clarify that the described limitations are generic to Simulacrum data, not specific to the package.

[ ] Remove if(!require("devtools")), and add a comment # before install.packages("devtools")

[ ] Make the 2016-2019 coverage more generic, explaining how the package works with earlier Simulacrum versions (v1.2.0).

[ ] Explain why the data restriction needs to be applied to the code before running on CAS data, or clarify if NDRS runs queries on full data.

[ ] Structure: Add a reference for researchers to look up the structure of the CAS database.

[ ] In line 96, add a linebreak before "Time Management".

[ ] Add a reference to the CAS data: https://doi.org/10.1093/ije/dyz076.

[ ] Consistently use either "the Simulacrum".

**Code Snippets & Examples**

[ ] Remove if(!require("devtools")), and add a comment # before install.packages("devtools")

[ ] In the code snippet for querying data, add a linebreak before ON to avoid line cut-off.

[ ] Review and edit create_workflow() examples in the paper, README, and vignette to ensure they create valid workflows.

[ ] In create_workflow(): Change the output data frame name from "data" to "query_result" (or similar) in the vignette.

[ ] Ensure consistency in variable names (e.g., query_result <- df1/data).

[ ] Implement replacement of Simulacrum table names with CAS table names (e.g., AV_PATIENT instead of SIM_AV_PATIENT).

[ ] In the README's create_workflow() example: Define sim_av_tumour if used in cancer_grouping().

[ ] Ensure the output of cancer_grouping() is assigned to an object, not just printed.

[ ] In the vignette's analysis section, address the GENDER.X issue by specifying the column to use in the SELECT statement with aliases (e.g., SELECT SIM_AV_PATIENT.GENDER gender, ...).

**Package Functionality**
[ ] Reconsider compute_time_limit(): Leave the interpretation of total execution time to the user.

[ ] Address the warning from read_simulacrum() (e.g., Warning message: In read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour")) :

[ ] Explain in the vignette why the warning appears.

**GitHub Repository Management**
[ ] Add bug report issue templates.

[ ] Add feature request issue templates.
