### 2025-05-22 026a7b2
changed validation to testing and changed access to real data to access to outputs of workflows
