+++ simulacrumWorkflowrRef.bib
@@ -32,7 +32,7 @@
   title = {{Guide to using the Simulacrum to support NDRS data requests}},
   organization = {National Disease Registration Service (NDRS)},
   year = {2023},
-  note = {Link: https://digital.nhs.uk/ndrs/data/data-outputs/cancer-publications-and-tools/simulacrum/simulacrum-user-guide} 
+  note = {Link: https://digital.nhs.uk/ndrs/data/data-outputs/cancer-publications-and-tools/simulacrum/simulacrum-user-guide/submitting-code-to-ndrs-for-a-data-release} 
 }
 
 @misc{nielsen2024simulacrum,
