+++ NAMESPACE
@@ -17,8 +17,8 @@ export(survival_days)
 import(dplyr)
 importFrom(DBI,dbConnect)
 importFrom(DBI,dbDisconnect)
+importFrom(DBI,dbGetQuery)
 importFrom(DBI,dbIsValid)
-importFrom(DBI,dbWriteTable)
 importFrom(RCurl,url.exists)
 importFrom(dplyr,select)
 importFrom(lubridate,as.duration)

[…]
