### 2025-05-25 CLINDA-AAU/simulacrumWorkflowR#1 @goldingn
Hi everyone, thanks for your enthusiastic reviews and responses!

It looks like everything is ticking along here, but please let me know if there's anything I can help with. E.g. if you want input on what does/doesn't need to be implemented to meet JOSS' acceptance criteria. I'll be away for the next week and a bit, but can respond on my return
