+++ simulacrumWorkflowrRef.bib
@@ -60,6 +60,21 @@
   note = {Link: https://digital.nhs.uk/ndrs/data/data-outputs/cancer-publications-and-tools/simulacrum/simulacrum-user-guide/submitting-code-to-ndrs-for-a-data-release} 
 }
 
+@online{odbc2025main,
+  author = {Microsoft},
+  title = {Microsoft Open Database Connectivity (ODBC)},
+  year = {2024}, 
+  url = {https://learn.microsoft.com/en-us/sql/odbc/microsoft-open-database-connectivity-odbc?view=sql-server-ver17},
+  urldate = {2025-05-26}
+}
+
+@online{oracle2025main,
+  author = {Oracle Corporation},
+  title = {Oracle},
+  year = {2025}, 
+  url = {https://www.oracle.com/},
+  urldate = {2025-05-26}
+}
 
 @Manual{Rsoftware2024stat,
  title = {R: A Language and Environment for Statistical Computing},
@@ -78,3 +93,14 @@
   howpublished = {\url{https://simulacrum.healthdatainsight.org.uk/}},
   note = {Accessed: 2025-05-23}
 }
+
+@online{sqlite2025main,
+  author = {SQLite},
+  title = {SQLite},
+  year = {2025}, 
+  url = {https://www.sqlite.org/},
+  urldate = {2025-05-26}
+}
+
+
+
