### 2025-05-26 d803599
added text for the vignette
