### 2025-05-27 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
Hi @tzuV Thanks for starting the check list, it looks like a good plan! Note that some of my points were truly meant as suggestions, and are not requirements. I would suggest to focus on the code snippets, and to explain the data coverage restrictions.
