### 2025-05-28 d9190ef
Edited paragraph 2 flow and fixed scattered typos
