### 2025-05-28 320d121
fixed NDRS misspellings (my own)
