+++ simulacrumWorkflowrRef.bib
@@ -40,7 +40,7 @@
   pages = {16--16h},
   year = {2020},
   doi = {10.1093/ije/dyz076},
-  month = feb,
+  month = {feb},
   note = {Published: 23 April 2019}
 }
 
