### 2025-06-04 f3e67e4
Fixed minor bug in paper.md
