### 2025-06-04 4380842
Fixed minor bug in paper.md
