+++ simulacrumWorkflowrRef.bib
@@ -54,6 +54,7 @@
 }
 
 @manual{national2023guide,
+  author = {NDRS}
   title = {Guide to using the Simulacrum to support NDRS data requests},
   organization = {National Disease Registration Service (NDRS)},
   year = {2023},
