### 2025-06-06 8580370
big fix in .bib
