### 2025-06-10 ae2e702
updated paper and .bib according to a new paper which has been published
