+++ simulacrumWorkflowrRef.bib
@@ -53,6 +53,16 @@
   note = {Accessed: 2025-05-23}
 }
 
+@article{kafatos2025syndata,
+  title = {Leveraging Synthetic Data to Facilitate Research: A Collaborative Model for Analyzing Sensitive National Cancer Registry Data in England},
+  author = {Kafatos, George and Levy, Julia and Jose, Sophie and Frayling, Lora and McInerney, Christopher J. and Moore, George and Johnson, Owen and Rashbass, Jem and Rous, Brian},
+  journal = {Therapeutic Innovation & Regulatory Science},
+  year = {2025},
+  month = {Jun},
+  doi = {10.1007/s43441-025-00820-z},
+  note = {Published: 05 June 2025}
+}
+
 @manual{national2023guide,
   author = {NDRS},
   title = {Guide to using the Simulacrum to support NDRS data requests},
