### 2025-06-13 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
@DrEspresso, I apologize for the extra week of waiting. I have made corrections according to all of your suggestions. After I made the suggestions, I passed them around to the co-authors, who agreed to the latest version of the package and paper. A list of all the corrections with responses can be found below:

**Paper/Documentation**

**Content & Clarity**
[X] Add references/links to Simulacrum, Cancer Analysis System (CAS), and Health Data Insight (HDI) where mentioned.

**Response**
References with embedded links have been added to the manuscript.

[X] Clarify "HDI" as "Health Data Insight" in the main text and add regional labels (e.g., "United Kingdom's National Disease Registration Service").

[X] In the summary, change "validating statistical analyses" to "testing" and/or "planning" statistical analyses.

**Response**
This has been changed to "planning statistical analysis."

[X] In line 30 & 43, clarify "releasing the data" to reflect that aggregated results are released, not the raw data.

**Response **

The previous version stated:
“*Scripts developed on the Simulacrum can be sent to NDRS for execution on the CAS database. This involves first adjusting the code (e.g., SQL queries need further processing by NDRS due to 1) structural differences between Simulacrum and CAS, 2) non-public details/specifications of the CAS database and 3) code alignment with NDRS best practices), executing the code, assessing outputs with respect to patient privacy and returning these outputs.*”

This has been changed to:

“*Scripts developed on the Simulacrum can be sent to NDRS for execution on the CAS database. This involves first adjusting the code (e.g., SQL queries need further processing by NDRS due to 1) structural differences between the Simulacrum and CAS, 2) non-public details/specifications of the CAS database and 3) code alignment with NDRS best practices), executing the code, assessing outputs with respect to patient privacy and returning aggregated outputs.*”

Additionally, we changed the line:

“*The process of accessing the real data through Simulacrum requires users to construct SQL queries for extracting data from the CAS database and R code for analysis.*”

to:

“*The process of having workflows executed on the real data through the Simulacrum requires users to construct SQL queries for extracting data from the CAS database and R code for analysis.*”

This change emphasizes that users cannot directly access private data.

[X] Add "as of 2025" to the sentence regarding the 3-hour free-of-charge limit.

**Response**
This has been added to the manuscript.

[X] In Figure 1: Remove the Simulacrum version unless it's a requirement (if so, add to Limitations and main text).

**Response**
To the best of our knowledge, there is no official documentation requiring the latest version of the Simulacrum. However, based on our communication with NDRS, we understand that the newest version is necessary because it most closely resembles the CAS database. I have added a paragraph to the main text to reflect this. I have also corrected the text to state that the analysis is executed on the CAS database.

[X] In figure 1: Swap "NHS" for "NDRS".

**Response**
This has been changed in the figure.

[X] In line 70, swap "NHS servers" with "CAS database servers".

**Response**
This has been implemented in the manuscript.

[X] Cite R and all R packages from the dependencies (e.g., using citation() in R).

**Response**
This has been implemented in the main text and the .bib file.

[X] Make the 2016-2019 coverage more generic, explaining how the package works with earlier Simulacrum versions (v1.2.0).
Response

We have added the following content to the coverage section:
“*- Coverage: The Simulacrum reflects diagnoses from 2016–2019, while CAS includes records dating back to 1971. The 2016-2019 restriction needs to be added to the code for running on CAS, as this time period is only provided in the free tier. Periods of the CAS data extending beyond The Simulacrum v2.1.0 will require a formal data release request and a cost estimate provided by DARS given the scope of data needed.*”

Additionally, the package is built around the newest version of the Simulacrum, and users are advised to use it, as this is the unofficial expectation for running an analysis on the CAS database. I have addressed this in the limitations section:

“*The newest version of the Simulacrum is strongly advised for implementing the workflow on the CAS database because it resembles the CAS database more than earlier versions of the Simulacrum. Thus, the functionalities of this package are built for Simulacrum v2.1.0, meaning some functions will not support earlier versions of the Simulacrum.*”

[X] Explain why the data restriction needs to be applied to the code before running on CAS data, or clarify if NDRS runs queries on full data.

**Response**
An additional explanation has been added in the limitations section under the coverage point. In short, the free tier for accessing CAS data is restricted to the 2016-2019 period; data beyond that is part of a paid tier.

[X] Structure: Add a reference for researchers to look up the structure of the CAS database.

**Response **
Currently, the best resources for understanding the CAS database structure are the Simulacrum and its resource paper. HDI previously offered a service called 'CASexplore,' but it appears to have been discontinued. However, we have added a reference to the NDRS homepage, which contains all the necessary resources.

[X] In line 96, add a linebreak before "Time Management".

**Response**
This has been added to the manuscript.

[X] Add a reference to the CAS data: https://doi.org/10.1093/ije/dyz076.

Response
**Thank you for highlighting this. It has been added to the manuscript and the .bib file.**

[X] Consistently use either "the Simulacrum" or "Simulacrum".

**Response**
Thank you for pointing this out. The text has been updated to consistently use "the Simulacrum."

[X] Remove if(!require("devtools")), and add a comment # before install.packages("devtools")

**Response**
This change has been made in the README file.

[X] In the code snippet for querying data, add a linebreak before ON to avoid line cut-off.

**Response**
This has been added to the manuscript.

[X] Review and edit create_workflow() examples in the paper, README, and vignette to ensure they create valid workflows.

**Response **
Thank you for this valuable input. We have updated the code in the paper, README, vignette, and the create_workflow() function to ensure a streamlined approach to creating functional workflow files.

[X] In create_workflow(): Change the output data frame name from "data" to "query_result" (or similar) in the vignette.

**Response**
Thank you for pointing this out. Consistency has been established across the paper, README, and vignette.

[X] Ensure consistency in variable names (e.g., query_result <- df1/data).

**Response**
Variable names have been made consistent.

[X] In create_workflow(): Implement replacement of Simulacrum table names with CAS table names (e.g., AV_PATIENT instead of SIM_AV_PATIENT).

**Response**
Thank you for the advice. This has been implemented in the code, ensuring that the create_workflow() function now removes the “_sim” prefix from all inputs.

[X] In the README's create_workflow() example: Define sim_av_tumour if used in cancer_grouping().

**Response**
This has been added to the code in the README.

[X] Ensure the output of cancer_grouping() is assigned to an object, not just printed.

**Response**
This has been added to the code in the paper.

[X] In the vignette's analysis section, address the GENDER.X issue by specifying the column to use in the SELECT statement with aliases (e.g., SELECT SIM_AV_PATIENT.GENDER gender, ...).

**Response**
Thank you for pointing this out. This has been implemented in the vignette file in line 75-81.

Package Functionality & Warnings
[X] Reconsider compute_time_limit(): Leave the interpretation of total execution time to the user.

**Response**
I have changed the text to make the user aware of the time limitation but have removed the ACCEPT/REJECT part of the message.

[X] Address the warning from read_simulacrum() (e.g., Warning message: In read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour")) :

**Response**
Explanations have been added in the vignette in line 38-43.

[X] Explain in the vignette why the warning appears.

**Response**
An explanation in the vignette has been added for each warning in line 83-85, 156-157, and 198-199.

GitHub Repository Management
[X] Add bug report issue templates. [X] Add feature request issue templates.

**Response**
Both issue templates have been added as instructed.
