### 2025-06-13 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
One thing I only noticed now, unfortunately, is the definition of your tests: Your test files do not test the package's functions, but instead duplicated code copied into these files.

The test files should really only contain the tests, and must not load any packages (this is handled in the testthat.R file). I recommend to have a look at the [testing chapter](https://r-pkgs.org/testing-basics.html#test-organisation) of Hadley Wickham's and Jennifer Bryan's R Package book.

I am sorry for realising this only now, a month after you added these tests! But I think you can quickly fix these files?
