### 2025-06-13 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
@DrEspresso , all good! I will make sure to cover it on Monday according to the R package book and give a notice here in the threat.
