### 2025-06-16 1a04cae
Added automatic testing
