### 2025-06-16 943c2cf
removed the outputs folders
