### 2025-06-16 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
@DrEspresso, thank you for bringing the testing of the functions to my attention. I realize it would be a maintenance nightmare with the previous system. It should be fixed now. Please let me know if there is anything else.
