### 2025-06-17 5ffe881
Add expect_warning for query_result variable name in query_sql()
