### 2025-06-17 a7cbeb0
Change to a sequence of message > warning > warning for compute_time_limit()
