### 2025-06-19 CLINDA-AAU/simulacrumWorkflowR#1 @DrEspresso
@tzuV I just found out how to generate the paper proof with the editorialbot, and had a look at the final pdf. There are three more minor things I noticed in the pdf:

- You still have `if(!require("devtools"))` in the paper. Why did you change it only in the README?
- In the "Loading data" section:
  - Instead of the generic `“/path/to/simulacrum/csv/files”` you now show a personal path.
  - The line `data_frames_list <- ...` runs out of the paper. Please add a linebreak after `read_simulacrum(dir,`
