### 2025-06-19 50d452e
corrected minor things in the paper
