### 2025-06-19 6e9608f
corrected minor things in the paper
