### 2025-07-10 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
@aghaynes, Just a quick follow-up on the paper. I am finalizing things before I go on vacation. Please let me know if there are any corrections or additions you would like me to address before I leave.
