### 2025-07-11 CLINDA-AAU/simulacrumWorkflowR#1 @aghaynes
Sorry, I've been swamped lately...

Thanks for addressing the points we've made. I don't think I have anything else.

I'll get on the pkgdown site for you, but I think that won't be until later in the summer. Expect a PR at some point 😄
