### 2025-07-11 CLINDA-AAU/simulacrumWorkflowR#1 @tzuV
Thank you @aghaynes and no stress regarding the pkgdown. I am just happy that you your offer still stands :)

@goldingn, i can see i should supply a DOI for a repo on Zenodo where i store a copy of this repo. Should i post in this thread or in the official review thread?
