### 2025-07-15 ec78809
update simulacrum and HDI references
