+++ simulacrumWorkflowrRef.bib
@@ -45,7 +45,7 @@
 }
 
 @misc{hdi2025main, 
-  author = {Health Data Insight},
+  author = {HDI},
   year = {2025},
   month = {05},
   title = {Health Data Insight CIC (HDI)},
@@ -97,7 +97,7 @@
 }
 
 @misc{simulacrum2025main, 
-  author = {simulacrum.healthdatainsight.org.uk},
+  author = {HDI},
   year = {2025},
   month = {05},
   title = {Simulacrum: Artificial patient-like cancer data to help researchers gain insights},
