### 2025-07-15 a2f2aa8
update oracle reference
