+++ simulacrumWorkflowrRef.bib
@@ -80,7 +80,7 @@
 }
 
 @online{oracle2025main,
-  author = {Oracle Corporation},
+  author = {{Oracle Corporation}},
   title = {Oracle},
   year = {2025}, 
   url = {https://www.oracle.com/},
