### 2025-07-27 #8120 @goldingn
Thank you @tzuV! It looks like there's just one issue remaining - editorialbot spotted a DOI you can use for the R citation (message above). Could you please add that to the bib file?

After this, the final step will be rerunning the check and recommending acceptance!
