### 2025-07-27 7e26994
Update simulacrumWorkflowrRef.bib
added doi supplied by the editorialbot from the JOSS review
