+++ simulacrumWorkflowrRef.bib
@@ -94,6 +94,7 @@
  address = {Vienna, Austria},
  year = {2024},
  url = {https://www.R-project.org/},
+ doi = {10.32614/r.manuals}
 }
 
 @misc{simulacrum2025main, 
