### 2025-07-29 #8120 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
