### 2025-08-09 #8120 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] versioning: your Zenodo deposit shows v2 and the version here (and recent tagged release on GH) shows v1 – these need to be consistent
- [ ] licensing: Zenodo seems to have picked up two licenses (MIT and Creative Commons), and in your repo there are two license files but one doesn't have any details, and the other is just an MIT license
- [ ] citations: check capitalization (e.g. need a `{England}` to force that to capitalize for Henson et al.)
- [ ] paper: please rewrite text immediately under "Key functionalities" so that it is a sentence rather than a fragment
- [ ] paper: it's a bit long, and the content of page 3 is really more suitable for package documentation than for the JOSS paper, I would remove that whole page
- [ ] paper: under Limitations, after bulleted list, Simulacrum should be capitalized
- [ ] paper, same paragraph: add commas after "well-balanced" and "comprehensive"
- [ ] paper, Time Management paragraph: I'm not sure what the "time Similarly" bit means
