### 2025-08-10 b51aad5
Correction of capitalizations in citations
