+++ simulacrumWorkflowrRef.bib
@@ -1,5 +1,5 @@
 @article{bullward2023research,
-  title = {Research Paper: Process Mining and Synthetic Health Data: Reflections and Lessons Learnt},
+  title = {Research Paper: {Process} {Mining} and {Synthetic} Health Data: Reflections and Lessons Learnt},
   author = {Bullward, A and Aljebreen, A and Coles, A and {Mclnerney}, C and Johnson, O},
   booktitle = {Process Mining Workshops},
   editor = {Montali, M and Senderovich, A and Weidlich, M},
@@ -9,30 +9,29 @@
   address = {Cham},
   year = {2023},
   doi = {10.1007/978-3-031-27815-0_25},
-  note = {ICPM 2022} 
+  note = {ICPM 2022}
 }
 
 
 @Manual{grothendieck2017sqldf,
- title = {sqldf: Manipulate R Data Frames Using SQL},
- author = {Grothendieck, G},
- year = {2017},
- note = {R package version 0.4.11},
- url = {https://CRAN.R-project.org/package=sqldf},
+  title = {{sqldf}: Manipulate {R} Data Frames Using {SQL}},
+  author = {Grothendieck, G},
+  year = {2017},
+  note = {R package version 0.4.11},
+  url = {https://CRAN.R-project.org/package=sqldf},
 }
-
 
 
 @manual{frayling2023simulacrum,
-  title = {Simulacrum v2 User Guide},
+  title = {{Simulacrum} v2 User Guide},
   author = {Frayling, L and Jose, S},
   organization = {Health Data Insight},
   year = {2023},
-  note = {Link: Simulacrum-v2-User-Guide.pdf} 
+  note = {Link: Simulacrum-v2-User-Guide.pdf}
 }
 
 @article{henson2020data,
-  title = {Data Resource Profile: National Cancer Registration Dataset in England},
+  title = {Data Resource Profile: National Cancer Registration Dataset in {England}},
   author = {Henson, Katherine E and Elliss-Brookes, Lucy and Coupland, Victoria H and Payne, Elsita and Vernon, Sally and Rous, Brian and Rashbass, Jem},
   journal = {International Journal of Epidemiology},
   volume = {49},
@@ -44,19 +43,19 @@
   note = {Published: 23 April 2019}
 }
 
-@misc{hdi2025main, 
+@misc{hdi2025main,
   author = {HDI},
   year = {2025},
   month = {05},
-  title = {Health Data Insight CIC (HDI)},
+  title = {Health Data Insight {CIC} ({HDI})},
   howpublished = {\url{https://healthdatainsight.org.uk/}},
   note = {Accessed: 2025-05-23}
 }
 
 @article{kafatos2025syndata,
-  title = {Leveraging Synthetic Data to Facilitate Research: A Collaborative Model for Analyzing Sensitive National Cancer Registry Data in England},
+  title = {Leveraging Synthetic Data to Facilitate Research: A Collaborative Model for Analyzing Sensitive National Cancer Registry Data in {England}},
   author = {Kafatos, George and Levy, Julia and Jose, Sophie and Frayling, Lora and McInerney, Christopher J. and Moore, George and Johnson, Owen and Rashbass, Jem and Rous, Brian},
-  journal = {Therapeutic Innovation & Regulatory Science},
+  journal = {Therapeutic Innovation \& Regulatory Science},
   year = {2025},
   month = {Jun},
   doi = {10.1007/s43441-025-00820-z},
@@ -65,54 +64,51 @@
 
 @manual{national2023guide,
   author = {NDRS},
-  title = {Guide to using the Simulacrum to support NDRS data requests},
-  organization = {National Disease Registration Service (NDRS)},
[…]
