### 2025-08-10 9040bb7
Edited paper.md based on reviewer comments. 1) rewrote the text under 'Key functionalities' to be less fragmented 2) removed the 'Workflow Illustration section 3) Capitalized simulacrum + added commas to same paragraph 4) Correct text under the limitations section regarding Time Management
