### 2025-08-10 6c768e3
Added two minor text corrections
