### 2025-08-12 6e5f09e
added JOSS acceptance status in readme.rmd
