+++ README.Rmd
@@ -15,6 +15,8 @@ library(knitr)
 
 # simulacrumWorkflowR
 
+[![status](https://joss.theoj.org/papers/10.21105/joss.08120/status.svg)](https://doi.org/10.21105/joss.08120)
+
 
 simulacrumWorkflowR is a package developed to assist users of the Simulacrum dataset in better preparing to use the dataset as a precursor to accessing real patient data in the Cancer Administration System (CAS).
 
