### 2025-08-12 54dc058
updated readme.qmd with the JOSS badge and changed the dir to the testing data
