+++ README.Rmd
@@ -68,7 +68,7 @@ The workflow is built around the sqldf package where the user are able to setup 
 The latest Simulacrum data is formatted identically to the real CAS data. Once downloaded, the read_simulacrum() function can automatically load the CSV files as data frames in R:
 
 ```{r}
-dir <- "C:/Users/p90j/Documents/simulacrum_v2.1.0/Data"
+dir <- system.file("extdata", "minisimulacrum", package = "simulacrumWorkflowR")
 # Automated data loading 
 data_frames_lists <- read_simulacrum(dir, selected_files = c("sim_av_patient", "sim_av_tumour")) 
 ```
